+++ LethalConstellations/Compat/NetworkThings.cs
@@ -16,6 +16,8 @@ namespace LethalConstellations.Compat
 
             LNetworkEvent consUnlockedSyncREQ = LNetworkEvent.Connect("consUnlockedSyncREQ", HostClientSync);
             LNetworkMessage<List<string>> constellationsUnlocked = LNetworkMessage<List<string>>.Connect("constellationsUnlocked", SyncUnlockHost, SyncUnlockClient);
+            LNetworkMessage<string> currentConst = LNetworkMessage<string>.Connect("currentConst", SyncCurrentHost, SyncCurrentClient);
+
 
         }
 
@@ -37,6 +39,25 @@ namespace LethalConstellations.Compat
                 constellationsUnlocked.SendServer(unlockedConst);
 
             //Collections.ConstellationsOTP = unlockedConst;
+        }
+
+        internal static void SyncCurrentSet(string currentConstellation)
+        {
+            if (!Plugin.instance.LethalNetworkAPI)
+                return;
+
+            if (currentConstellation.Length == 0)
+            {
+                Plugin.WARNING("Attempting to sync blank currentConstellation!!!");
+            }
+
+            LNetworkMessage<string> currentConst = LNetworkMessage<string>.Connect("currentConst", SyncCurrentHost, SyncCurrentClient);
+
+            if (LNetworkUtils.IsHostOrServer)
+                currentConst.SendClients(currentConstellation, LNetworkUtils.AllConnectedClients);
+            else
+                currentConst.SendServer(currentConstellation);
+
         }
 
         internal static void RequestSyncFromHost()
@@ -64,6 +85,7 @@ namespace LethalConstellations.Compat
             Plugin.Spam("Host sending collection");
 
             SyncUnlockSet(Collections.ConstellationsOTP);
+            SyncCurrentSet(Collections.CurrentConstellation);
         }
 
         internal static void SyncUnlockHost(List<string> newValue, ulong clientSending)
@@ -100,6 +122,40 @@ namespace LethalConstellations.Compat
             Plugin.Spam($"SYNCING NEW UNLOCK LIST");
             Collections.ConstellationsOTP = newValue;
             ClassMapper.UpdateConstellationUnlocks();
+        }
+
+        internal static void SyncCurrentHost(string newValue, ulong clientSending)
+        {
+            if (!Plugin.instance.LethalNetworkAPI)
+                return;
+
+            //Plugin.Spam("Host received new list");
+            if (Collections.CurrentConstellation == newValue)
+                return;
+
+            Plugin.Spam($"SYNCING NEW UNLOCK LIST - HOST");
+            Collections.CurrentConstellation = newValue;
+            ClassMapper.UpdateConstellationUnlocks();
+            SaveManager.SaveCurrentConstellation(Collections.CurrentConstellation);
+        }
+
+        internal static void SyncCurrentClient(string newValue)
+        {
+            if (!Plugin.instance.LethalNetworkAPI)
+                return;
+
+            //Plugin.Spam("Sync Called");
+
+            if (LNetworkUtils.IsHostOrServer)
+                return;
+
+            //Plugin.Spam("Not host");
+
+            if (Collections.CurrentConstellation == newValue)
+                return;
+
+            Plugin.Spam($"SYNCING NEW UNLOCK LIST");
+            Collections.CurrentConstellation = newValue;
         }
     }
 }