+++ LethalConstellations/PluginCore/Plugin.cs
@@ -21,7 +21,7 @@ namespace LethalConstellations
         {
             public const string PLUGIN_GUID = "com.github.darmuh.LethalConstellations";
             public const string PLUGIN_NAME = "LethalConstellations";
-            public const string PLUGIN_VERSION = "0.2.2";
+            public const string PLUGIN_VERSION = "0.2.3";
         }
         
         internal static ManualLogSource Log;